Streaming: Single retransmit timer per connection (ticket #2715) Only instantiate ResendPacketEvent for fast retransmit Move packet timeout field from ResendPacketEvent to PacketLocal Set window size to 1 on timeout Always adjust window when acked, even if the packet was sent more than once Reduce INITIAL_WINDOW_SIZE from 6 to 3 to reduce router DH overhead Set maximum burst retransmissions to 16 Closer compliance to RFCs 5681 and 6298 Prep for additional changes (Westwood) Other minor cleanups Original patch from zlatinb

+++ apps/streaming/java/src/net/i2p/client/streaming/impl/TCBShare.java
@@ -40,7 +40,7 @@ class TCBShare {
     private static final String WDW_DAMP_PROP="i2p.streaming.tcbcache.wdwDampening";
     private static final String RTTDEV_DAMP_PROP="i2p.streaming.tcbcache.rttdevDampening";
     /////
-    private static final int MAX_RTT = ((int) Connection.MAX_RESEND_DELAY) / 2;
+    private static final int MAX_RTT = Connection.MAX_RESEND_DELAY / 2;
     private static final int MAX_RTT_DEV = (int) (MAX_RTT * 1.5);
     private static final int MAX_WINDOW_SIZE = ConnectionPacketHandler.MAX_SLOW_START_WINDOW;
